Adjust V-Kit to the Connect oracle changes
We are about to roll out the Connect oracle on Mezo testnet. We need
to ensure existing validators can seamlessly connect with their
Connect sidecars once the upgrade happens.

Pull request https://github.com/mezo-org/mezod/pull/359 introduces the
following on the `mezod` side:
- The `init` command now adds the oracle-specific config to `app.toml`.
- The `start` command exposes oracle-specific flags (`--oracle.*` family)
  with default values.
- The `customize_configuration` of the `entrypoint.sh` enables the oracle
  by default and allows setting its address using the
  `MEZOD_ORACLE_ORACLE_ADDRESS` env (naming adjusted to the `app.toml` param
  according to Vyper's convention). If this env is not set, the default
  value of `connect-sidecar:8080` is used.

To make everything work, we are introducing some adjustments to the following
variants of the V-Kit:
- Helm: We are setting the `MEZOD_ORACLE_ORACLE_ADDRESS: "localhost:8080"`
  env. This is necessary to pass the correct address to the `mezod` entrypoint.
  Otherwise, `connect-sidecar:8080` will be used by default which is wrong
  in the Kubernetes context
- Native: We are exporting `MEZOD_ORACLE_ORACLE_ADDRESS` and
  `MEZOD_ORACLE_ENABLED` with appropriate values. This is not strictly needed
  due to `mezod` default values of the `--oracle.*` flags but, we are
  doing it for consistency and because the Connect sidecar port is
  configurable in this variant of the V-Kit

Docker variant doesn't have to be adjusted because it performs `init` on
restart. That will add the oracle-specific config to `app.toml` and
leverage the default oracle address set in the entrypoint which is
correct in the context of Docker.