Adding rpc-url optional flag description
For some validators a default localnode url won't work to submit their
application. It can be fixed by adding --rpc-url <url> flag pointing to
a valid RPC url. This info should also be included in the --help of the
command itself, but for the time being as we are onboarding more
validators, we can add this info in the docs here.