Pipe keyring password to Ethereum sidecar
Ethereum sidecar recently started to require the keyring and a key it can use to execute bridge out attestations. However, if the keyring is `file`, the Ethereum sidecar process prompts for keyring passport several times upon startup. Here we solve that by piping the keyring password through `yes`. We are also setting the home dir wherever missing as it is required to properly extract the keyring.